restructure dav module in preparation for google port
